+++ transport/grpc/inbound_test.go
@@ -45,3 +45,23 @@ func TestInboundMechanics(t *testing.T) {
 	assert.NoError(t, inbound.Stop())
 	assert.Nil(t, inbound.Addr())
 }
+
+func TestInboundIntrospection(t *testing.T) {
+	listener, err := net.Listen("tcp", "127.0.0.1:0")
+	require.NoError(t, err)
+	inbound := NewTransport().NewInbound(listener)
+	inbound.SetRouter(newTestRouter(nil))
+
+	assert.Equal(t, transportName, inbound.Introspect().Transport, "unexpected transport name")
+	assert.Equal(t, "Stopped", inbound.Introspect().State, "expected 'Stopped' state")
+	assert.Empty(t, inbound.Introspect().Endpoint, "unexpected endpoint")
+
+	require.NoError(t, inbound.Start())
+	assert.Equal(t, "Started", inbound.Introspect().State, "expected 'Started' state")
+	assert.NotEmpty(t, inbound.Introspect().Endpoint)
+	assert.Equal(t, inbound.Addr().String(), inbound.Introspect().Endpoint, "unexpected endpoint")
+
+	assert.NoError(t, inbound.Stop())
+	assert.Equal(t, "Stopped", inbound.Introspect().State, "expected 'Stopped' state")
+	assert.Empty(t, inbound.Introspect().Endpoint, "unexpected endpoint")
+}